+++ TouchMVP/app/src/main/java/com/example/touchmvp/JoyStick.java
@@ -2,19 +2,19 @@ package com.example.touchmvp;
 
 /** 모델 **/
 public class JoyStick {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    JoyStick(int x, int y){
+    JoyStick(float x, float y){
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 }
